entrypoint/pyproject.toml
[tool.poetry]
name = "entrypoint"
version = "0.1.0"
description = ""
authors = ["Michael Buehlmann <buehlmann.michi@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
flows-entry-test = "entrypoint.scripts.test_script:cli"
setup-entry-fct = "entrypoint.scripts.setup_entrypoint:cli"

[tool.poetry.dependencies]
python = "^3.11"
globus-compute-sdk = "^2.24.0"
python-dotenv = "^1.0.1"
datasets = { path = "../datasets", develop = true }
click = "^8.1.7"
globus-compute-endpoint = "^2.24.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"

[tool.mypy]
plugins = ['pydantic.mypy']
disable_error_code = ['import-untyped', 'import-not-found']


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
